+++ CozyPlayer/CHANGELOG.md
@@ -1,6 +1,19 @@
 # Cozy Player ChangeLog
 
-## Version 1.x
+## Version 1.2.x
+
+### 1.2.0
+
+* Added more cases where the chat speaker name can be clicked
+* **"Show target on rolls"**: option changed to **"Targets: Add to chat"**. 
+* **"Targets: Add to chat"**: The targets on the chat message shows to every other player, not only to the sender.
+* **"Targets: Add to chat"**: supports most of Better Rolls rolls if set to "implicit"
+* **"Targets: Add to chat"** now have a "On any message" option
+* **Tooltip**: Added an option to select the tooltip style between black and white
+* **Tooltip**: Added an option to show consumable items
+* **Door Interaction** Feature was removed from this module. I've created a separated module to this feature: Arms Reach.
+
+## Version 1.1.x
 
 ### 1.1.1